requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pyyaml==6.0.1
requests==2.31.0
SQLAlchemy==2.0.23
